requirements.txt
fastapiuvicornscikit-learnpandasjoblibnumpy